refactor(polling): split polling views and modals
