+++ bot/features/polling/modals/__init__.py
@@ -1,0 +1,9 @@
+"""Polling modal components."""
+
+from .add_option import AddOptionModal
+from .create_poll import PollInputModal
+
+__all__ = [
+    "AddOptionModal",
+    "PollInputModal",
+]
