+++ bot/features/polling/views/__init__.py
@@ -1,0 +1,17 @@
+"""Polling Discord view components.
+
+The modal classes are re-exported here for compatibility with existing
+``bot.features.polling.views`` imports.
+"""
+
+from .poll_view import PollView
+from bot.features.polling.modals import (
+    AddOptionModal,
+    PollInputModal,
+)
+
+__all__ = [
+    "AddOptionModal",
+    "PollInputModal",
+    "PollView",
+]
